requirements.txt
Django==5.2.10
mysqlclient==2.2.7
groq==0.15.0
python-dotenv==1.2.1
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
djangorestframework==3.15.2
openpyxl==3.1.5
xhtml2pdf==0.2.17
reportlab==4.4.9
